json handle

+++ agent/jvm/jvminfo/memory.go
@@ -7,7 +7,7 @@ import (
 type Memory struct {
 	//Request interface{} `json:"request"`
 	Value memoryValue `json:"value"`
-	TimeStamp int32 `json:"timestamp"`
+	TimeStamp int64 `json:"timestamp"`
 	Status int32 `json:"status"`
 }
 type memoryusage struct {
